pyproject.toml
[project]
name = "catalog-research-agent"
version = "0.1.0"
description = "Catalog Research Agent for software lifecycle research automation"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typing-extensions>=4.0.0",
    "pydantic>=2.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
